Adds option to draw a line between 2d points
This is a feature used in MAXIV for some of the custom tasks.
It adds an object type in context menu for a pair of 2D Points.
Existing `HELICAL` object type did not work, as it works for a pair of
3D Point.
I could alternatively change context menu code, so that `HELICAL` object
type would be common across 2D and 3D point, I am not sure if such
behaviour would be desirable though.